requirements.txt
anthropic>=0.5.0
gitpython>=3.1.30
pyyaml>=6.0
click>=8.1.3
ansible-lint>=6.14.0
ruamel.yaml>=0.17.21
jinja2>=3.1.2
flask>=2.3.0
flask-wtf>=1.1.1
flask-bootstrap4>=4.0.2
werkzeug>=2.3.0
wtforms>=3.0.0
requests>=2.31.0
gunicorn>=20.1.0
